Patch 08/10 Vector indices are cast to unsigned long so that the compiler finds the right overloaded operator[].

+++ src/vector/cl_SV_copy.cc
@@ -22,7 +22,7 @@ const cl_SV_any copy (const cl_SV_any& src)
 	hv->type = src.pointer_type();
 	new (&hv->v) cl_SV_inner<cl_gcobject> (len);
 	for (std::size_t i = 0; i < len; i++)
-		init1(cl_gcobject, hv->v[i]) (src[i]);
+		init1(cl_gcobject, hv->v[(unsigned long)i]) (src[(unsigned long)i]);
 	return hv;
 }
